Seperated tests into files
