+++ src/NServiceBus.AzureFunctions.Analyzer.Tests/AzureFunctionsConfigurationAnalyzerTests.cs
@@ -153,263 +153,5 @@ class Foo
 
             return Assert(AzureFunctionsDiagnostics.OverrideLocalAddressNotAllowedId, source);
         }
-
-        [Test]
-        public Task DiagnosticIsReportedForRouteReplyToThisInstance()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar()
-    {{
-        var replyOptions = new ReplyOptions();
-        [|replyOptions.RouteReplyToThisInstance()|];
-
-        var sendOptions = new SendOptions();
-        [|sendOptions.RouteReplyToThisInstance()|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.RouteReplyToThisInstanceNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForRouteToThisInstance()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar()
-    {{
-        var options = new SendOptions();
-        [|options.RouteToThisInstance()|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.RouteToThisInstanceNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForRouteReplyToAnyInstance()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar()
-    {{
-        var options = new SendOptions();
-        [|options.RouteReplyToAnyInstance()|];
-
-        var replyOptions = new ReplyOptions();
-        [|replyOptions.RouteReplyToAnyInstance()|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.RouteReplyToAnyInstanceNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForPrefetchCount()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(ServiceBusTriggeredEndpointConfiguration endpointConfig)
-    {{
-        [|endpointConfig.Transport.PrefetchCount|] = 5;
-
-        var transportConfig = endpointConfig.Transport;
-        [|transportConfig.PrefetchCount|] = 5;
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.PrefetchCountNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForPrefetchCountAsExtension()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(TransportExtensions<AzureServiceBusTransport> transportExtension)
-    {{
-        [|transportExtension.PrefetchCount(5)|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.PrefetchCountNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForPrefetchMultiplier()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(ServiceBusTriggeredEndpointConfiguration endpointConfig)
-    {{
-        [|endpointConfig.Transport.PrefetchMultiplier|] = 5;
-
-        var transportConfig = endpointConfig.Transport;
-        [|transportConfig.PrefetchMultiplier|] = 5;
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.PrefetchMultiplierNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForPrefetchMultiplierAsExtension()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(TransportExtensions<AzureServiceBusTransport> transportExtension)
-    {{
-        [|transportExtension.PrefetchMultiplier(5)|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.PrefetchMultiplierNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForMaxAutoLockRenewalDuration()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(ServiceBusTriggeredEndpointConfiguration endpointConfig)
-    {{
-        [|endpointConfig.Transport.MaxAutoLockRenewalDuration|] = new System.TimeSpan(0, 0, 5, 0);
-
-        var transportConfig = endpointConfig.Transport;
-        [|transportConfig.MaxAutoLockRenewalDuration|] = new System.TimeSpan(0, 0, 5, 0);
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.MaxAutoLockRenewalDurationNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForMaxAutoLockRenewalDurationAsExtension()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(TransportExtensions<AzureServiceBusTransport> transportExtension)
-    {{
-        [|transportExtension.MaxAutoLockRenewalDuration(new System.TimeSpan(0, 0, 5, 0))|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.MaxAutoLockRenewalDurationNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForTimeToWaitBeforeTriggeringCircuitBreaker()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(ServiceBusTriggeredEndpointConfiguration endpointConfig)
-    {{
-        [|endpointConfig.Transport.TimeToWaitBeforeTriggeringCircuitBreaker|] = new System.TimeSpan(0, 0, 5, 0);
-
-        var transportConfig = endpointConfig.Transport;
-        [|transportConfig.TimeToWaitBeforeTriggeringCircuitBreaker|] = new System.TimeSpan(0, 0, 5, 0);
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.TimeToWaitBeforeTriggeringCircuitBreakerNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForTimeToWaitBeforeTriggeringCircuitBreakerAsExtension()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(TransportExtensions<AzureServiceBusTransport> transportExtension)
-    {{
-        [|transportExtension.TimeToWaitBeforeTriggeringCircuitBreaker(new System.TimeSpan(0, 0, 5, 0))|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.TimeToWaitBeforeTriggeringCircuitBreakerNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForEntityMaximumSize()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(ServiceBusTriggeredEndpointConfiguration endpointConfig)
-    {{
-        [|endpointConfig.Transport.EntityMaximumSize|] = 5;
-
-        var transportConfig = endpointConfig.Transport;
-        [|transportConfig.EntityMaximumSize|] = 5;
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.EntityMaximumSizeNotAllowedId, source);
-        }
-
-        [Test]
-        public Task DiagnosticIsReportedForEntityMaximumSizeAsExtension()
-        {
-            var source =
-                $@"using NServiceBus; 
-using System;
-using System.Threading.Tasks; 
-class Foo
-{{
-    void Bar(TransportExtensions<AzureServiceBusTransport> transportExtension)
-    {{
-        [|transportExtension.EntityMaximumSize(1)|];
-    }}
-}}";
-
-            return Assert(AzureFunctionsDiagnostics.EntityMaximumSizeNotAllowedId, source);
-        }
     }
 }
